+++ app/code/Deity/CatalogApi/Api/ProductRepositoryInterface.php
@@ -1,0 +1,20 @@
+<?php
+declare(strict_types=1);
+
+namespace Deity\CatalogApi\Api;
+
+interface ProductRepositoryInterface extends \Magento\Catalog\Api\ProductRepositoryInterface
+{
+    /**
+     * Get product list
+     * @param \Magento\Framework\Api\SearchCriteriaInterface $searchCriteria
+     * @param bool $includeSubcategories
+     * @param mixed $withAttributeFilters
+     * @return \Deity\CatalogApi\Api\Data\ProductSearchResultsInterface
+     */
+    public function getList(
+        \Magento\Framework\Api\SearchCriteriaInterface $searchCriteria,
+        $includeSubcategories = false,
+        $withAttributeFilters = []
+    );
+}
