+++ app/code/Deity/CatalogApi/etc/webapi.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 <routes xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Webapi:etc/webapi.xsd">
-    <!--<route url="/V1/url" method="GET">
-        <service class="Deity\UrlRewriteApi\Api\GetUrlRewriteInterface" method="execute"/>
+    <route url="/V1/products" method="GET">
+        <service class="Deity\CatalogApi\Api\ProductRepositoryInterface" method="getList"/>
         <resources>
-            <resource ref="anonymous"/>
+            <resource ref="Magento_Catalog::products" />
         </resources>
-    </route>-->
+    </route>
 </routes>
