--- app/code/Deity/MagentoApi/Api/Catalog/AttributeManagementInterface.php
@@ -1,10 +1,0 @@
-<?php
-namespace Deity\MagentoApi\Api\Catalog;
-
-interface AttributeManagementInterface
-{
-    /**
-     * @return \Deity\MagentoApi\Api\Data\FilterInterface[]
-     */
-    public function getCategoryFilters();
-}
