--- app/code/Deity/MagentoApi/Api/ProductRepositoryInterface.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace Deity\MagentoApi\Api;
-
-interface ProductRepositoryInterface extends \Magento\Catalog\Api\ProductRepositoryInterface
-{
-    /**
-     * Get product list
-     * @param \Magento\Framework\Api\SearchCriteriaInterface $searchCriteria
-     * @param bool $includeSubcategories
-     * @param mixed $withAttributeFilters
-     * @return \Deity\MagentoApi\Api\Data\ProductSearchResultsInterface
-     */
-    public function getList(\Magento\Framework\Api\SearchCriteriaInterface $searchCriteria, $includeSubcategories = false, $withAttributeFilters = []);
-}
