+++ app/code/Deity/MagentoApi/etc/di.xml
@@ -2,13 +2,8 @@
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
 
     <!-- INTERFACE PREFERENCES FOR DI -->
-    <preference for="Deity\MagentoApi\Api\Catalog\AttributeManagementInterface" type="Deity\MagentoApi\Model\Catalog\AttributeManagement" />
-    <preference for="Deity\MagentoApi\Api\CategoryManagementInterface" type="Deity\MagentoApi\Model\CategoryManagement" />
     <preference for="Deity\MagentoApi\Api\Customer\AddressRepositoryInterface" type="Deity\MagentoApi\Model\Customer\AddressRepository" />
     <preference for="Deity\MagentoApi\Api\ContactFormInterface" type="Deity\MagentoApi\Model\ContactForm" />
-    <preference for="Deity\MagentoApi\Api\Data\CategoryTreeInterface" type="Deity\MagentoApi\Model\Category" />
-    <preference for="Deity\MagentoApi\Api\Data\FilterInterface" type="Deity\MagentoApi\Model\Api\Data\Filter" />
-    <preference for="Deity\MagentoApi\Api\Data\FilterOptionInterface" type="Deity\MagentoApi\Model\Api\Data\FilterOption" />
     <preference for="Deity\MagentoApi\Api\Data\GalleryMediaEntrySizeInterface" type="Deity\MagentoApi\Model\GalleryMediaEntrySize" />
     <preference for="Deity\MagentoApi\Api\Data\InfoDataInterface" type="Deity\MagentoApi\Model\InfoData" />
     <preference for="Deity\MagentoApi\Api\Data\MenuInterface" type="Deity\MagentoApi\Model\Api\Data\Menu" />
@@ -25,7 +20,6 @@
     <preference for="Deity\MagentoApi\Api\Payment\Data\PaypalDataInterface" type="Deity\MagentoApi\Model\Payment\PaypalData" />
     <preference for="Deity\MagentoApi\Api\Payment\GuestPaypalInterface" type="Deity\MagentoApi\Model\Payment\GuestPaypal" />
     <preference for="Deity\MagentoApi\Api\Payment\PaypalInterface" type="Deity\MagentoApi\Model\Payment\Paypal" />
-    <preference for="Deity\MagentoApi\Api\ProductRepositoryInterface" type="Deity\MagentoApi\Model\ProductRepository" />
     <preference for="Deity\MagentoApi\Api\QuoteMaskInterface" type="Deity\MagentoApi\Model\QuoteMask" />
 
     <!-- OVERRIDES -->
@@ -34,24 +28,6 @@
     <preference for="Magento\Paypal\Model\Express" type="Deity\MagentoApi\Model\Payment\Paypal\Express" />
     <preference for="Magento\Paypal\Model\Express\Checkout" type="Deity\MagentoApi\Model\Payment\Paypal\Express\Checkout" />
     <preference for="Magento\Paypal\Model\Pro" type="Deity\MagentoApi\Model\Payment\Paypal\Pro" />
-
-    <!-- ARGUMENTS FOR CONSTRUCTORS -->
-    <type name="Magento\Catalog\Model\ProductRepository">
-        <arguments>
-            <argument name="collectionFactory" xsi:type="object">ProductRepositoryCollectionFactory</argument>
-            <argument name="searchResultsFactory" xsi:type="object">ProductRepositorySearchResultsFactory</argument>
-        </arguments>
-    </type>
-    <virtualType name="ProductRepositoryCollectionFactory" type="Magento\Catalog\Model\ResourceModel\Product\CollectionFactory">
-        <arguments>
-            <argument name="instanceName" xsi:type="string">Deity\MagentoApi\Model\ResourceModel\Product\Collection</argument>
-        </arguments>
-    </virtualType>
-    <virtualType name="ProductRepositorySearchResultsFactory" type="Magento\Catalog\Api\Data\ProductSearchResultsInterfaceFactory">
-        <arguments>
-            <argument name="instanceName" xsi:type="string">Deity\MagentoApi\Api\SearchResults</argument>
-        </arguments>
-    </virtualType>
 
     <!-- PLUGINS -->
     <type name="Magento\Sales\Api\OrderManagementInterface">
